adding some gentypes
