+++ eyes2-lib/src/entity/genotype/genotype.rs
@@ -1,9 +1,13 @@
+use crate::entity::update::UpdateQueue;
+use crate::entity::Creature;
+use crate::Settings;
+
 pub trait Genotype
 where
     Self: Sized,
 {
     // constructor
-    fn new(entity: Entity, config: Settings) -> Self;
+    fn new(creature: Creature, config: Settings) -> Self;
 
     // execute the next instruction of your Genomic code
     fn tick(&mut self, queue: &mut UpdateQueue);
