+++ eyes2-lib/src/entity/genotype/random.rs
@@ -7,27 +7,13 @@
 const GENOME: usize = 1000;
 
 #[derive(Debug)]
-pub struct Processor {
-    pub energy: i32,       // energy level
-    ip: u16,               // instruction pointer
-    a: u16,                // accumulator
-    i: [u16; 5],           // registers
-    breed_rate: u32,       // rate at which the creature breeds
-    mutation_rate: u32,    // rate at which the creature mutates
-    genome: [u16; GENOME], // the genome i.e. the instructions to be executed
+pub struct RandomGenotype {
+    pub energy: i32, // energy level
 }
 
-impl Processor {
-    pub fn new(energy: i32) -> Processor {
-        Processor {
-            energy,
-            ip: 0,
-            a: 0,
-            i: [0; 5],
-            breed_rate: 0,
-            mutation_rate: 0,
-            genome: Processor::randomize(),
-        }
+impl RandomGenotype {
+    pub fn new(energy: i32) -> RandomGenotype {
+        RandomGenotype { energy }
     }
 
     pub fn randomize() -> [u16; GENOME] {
@@ -39,14 +25,5 @@ impl Processor {
     }
 
     // totally dummy instruction set for now
-    pub fn tick(&mut self) {
-        self.ip = (self.ip + 1) % (GENOME as u16);
-        let instruction = self.genome[self.ip as usize];
-        match instruction {
-            0 => self.a = self.i[0],
-            1 => self.a = self.breed_rate as u16,
-            2 => self.a = self.mutation_rate as u16,
-            _ => (),
-        }
-    }
+    pub fn tick(&mut self) {}
 }
